% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsta_data}
\alias{gsta_data}
\title{CRUTEM World Mean Temperature Data Set from 1851 to 2022}
\format{
\subsection{\code{gsta_data}}{

A data set with 172 rows and 4 columns, for Jan, Apr, Jul and Oct from 1851-2022
\describe{
\item{Jan}{world mean temperature for January}
\item{Apr}{world mean temperature for April}
\item{Jul}{world mean temperature for July}
\item{Oct}{world mean temperature for October}
}
}
}
\source{
\url{https://crudata.uea.ac.uk/cru/data/temperature/CRUTEM.5.0.1.0.stat4post.txt.gz}
}
\usage{
gsta_data
}
\description{
CRUTEM World Mean Temperature Data Set from 1851 to 2022
}
\keyword{datasets}
